testing the render props and context api, wip
