+++ src/auth-context.js
@@ -1,0 +1,13 @@
+import React from 'react'
+
+const authContext = React.createContext({
+  clientId : "",
+  provider : "",
+  clientSecret : "",
+  pkce : false,
+  authenticate : () => {
+
+  }
+})
+
+export default authContext;
