+++ src/index(legacy).js
@@ -1,0 +1,5 @@
+import { AuthContext } from './components/auth-context'
+import { authenticated } from './authenticated'
+import { getLocalToken } from './local-token'
+
+export { AuthContext, authenticated, getLocalToken }
