+++ src/index.js
@@ -1,5 +1,5 @@
-import { AuthContext } from './components/auth-context'
-import { authenticated } from './authenticated'
+import AuthContext  from './auth-context'
+import authenticated  from './authenticated'
 import { getLocalToken } from './local-token'
 
-export { AuthContext, authenticated, getLocalToken }
+export {AuthContext,authenticated,getLocalToken}
